Adds parameter `keepOrder` to `RemoveObject[s]()` methods.
The current implementation uses `set`s to quickly remove a bunch of
objects. Unfortunately that can change the order of the remaining
objects.

This adds the parameter `keepOrder` to all `RemoveObject()` and
`RemoveObjects()` method and the initialization of `BatchedUpdate`.

By default this parameter is `False` to keep the original behaviour.